+++ test/SpooVault.test.cjs
@@ -1,5 +1,6 @@
 const { expect } = require("chai");
 const { ethers } = require("hardhat");
+const { deploySpooVault } = require("./helpers/deploySpooVault.cjs");
 
 describe("SpooVault EVM Contract Unit Tests", function () {
   let spooVault;
@@ -11,9 +12,7 @@ describe("SpooVault EVM Contract Unit Tests", function () {
   beforeEach(async function () {
     [owner, guardian1, guardian2, beneficiary] = await ethers.getSigners();
 
-    const SpooVault = await ethers.getContractFactory("SpooVault");
-    spooVault = await SpooVault.deploy();
-    await spooVault.waitForDeployment();
+    spooVault = await deploySpooVault();
   });
 
   describe("Public Key Registry", function () {